src/processing/requirements.txt
aiohttp>=3.7.4
async-timeout==3.0.1
attrs==21.2.0
azure-core==1.18.0
azure-functions==1.7.2
azure-storage-queue==12.1.6
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cryptography==3.4.8
idna==3.2
isodate==0.6.0
msrest==0.6.21
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
pycparser==2.20
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
